[MB-639] cast long string formats to long

+++ Airship/Inbox/UAInboxMessageList.m
@@ -207,7 +207,7 @@ NSString * const UAInboxMessageListUpdatedNotification = @"com.urbanairship.noti
                 }
             }];
         default:
-            UA_LWARN(@"Unable to perform batch update with invalid command type: %ld", command);
+            UA_LWARN(@"Unable to perform batch update with invalid command type: %ld", (long)command);
             return nil;
     }
 }
@@ -234,7 +234,7 @@ NSString * const UAInboxMessageListUpdatedNotification = @"com.urbanairship.noti
                 }
             }];
         default:
-            UA_LWARN(@"Unable to perform batch update with invalid command type: %ld", command);
+            UA_LWARN(@"Unable to perform batch update with invalid command type: %ld",(long)command);
             return nil;
     }
 }
